requirements.txt
-f /cvmfs/soft.computecanada.ca/custom/python/wheelhouse/avx
-f /cvmfs/soft.computecanada.ca/custom/python/wheelhouse/generic
absl-py==0.7.1
argh==0.26.2
astor==0.8.1
backcall==0.1.0
cachetools==4.0.0
certifi==2019.11.28
cftime==1.0.4.2
chardet==3.0.4
Click==7.0
configparser==4.0.2
cycler==0.10.0
Cython==0.29.14
decorator==4.4.1
docker-pycreds==0.4.0
gast==0.2.2
gitdb2==3.0.2
GitPython==3.0.8
google-auth==1.11.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
gql==0.2.0
graphql-core==1.1
grpcio==1.25.0
h5netcdf==0.7.4
h5py==2.9.0
idna==2.8
ipykernel==5.1.3
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.16.0
joblib==0.14.1
jupyter-client==5.3.4
jupyter-core==4.6.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
line-profiler==2.1.1
lz4==3.0.2
Markdown==3.1.1
matplotlib==3.1.1
netCDF4==1.5.3
numpy==1.17.4
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencv-python==3.4.4.19
opt-einsum==2.3.2
pandas==0.25.3
parso==0.6.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
promise==2.3
prompt-toolkit==3.0.2
protobuf==3.11.2
psutil==5.6.5
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.5.2
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.1.2
pyzmq==18.1.0
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
sentry-sdk==0.14.1
shortuuid==0.5.0
six==1.14.0
sklearn==0.0
smmap2==2.0.5
subprocess32==3.5.4
tensorboard==2.0.2
tensorflow-estimator==2.0.1
tensorflow-gpu==2.0.0
termcolor==1.1.0
tornado==6.0.3
tqdm==4.43.0
traitlets==4.3.3
urllib3==1.25.8
wandb==0.8.27
watchdog==0.9.0
wcwidth==0.1.8
Werkzeug==0.16.0
wrapt==1.11.2
